Fixed problem with exit status
The exit status (given with system call exit) was not correctly
transfered to parent step/process (using system call waitpid)

The exit status of each step is now also stored in the
skip_processes(_new).txt files.

The chmod system call was not working because it was not
operating on the 'mapped' file.

Added command line option for specifying the file to
log all output will be send to using the '-l' option
followed by the filename. Some of the executables also
print output to stdout and stderr, causing the output of
the program getting 'mixed' with the output of the program.

Fixed some problems found with -Wall including some
'bug' in the memory access code that might only show up
at some rare occasions.